Simple first refactor extracting the reconcilation report reshaping
